Remove model.physical package
