+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/model/physical/TableMetaDataTest.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.infra.metadata.schema.model.physical;
 
+import org.apache.shardingsphere.infra.metadata.schema.model.PhysicalColumnMetaData;
+import org.apache.shardingsphere.infra.metadata.schema.model.PhysicalTableMetaData;
 import org.junit.Before;
 import org.junit.Test;
